Move createIFrameUrl into MakeCodeFrame
